feat(BaseConfigurationRazor): can select home geofence via map

+++ TeslaSolarCharger/Shared/Dtos/BaseConfiguration/BaseConfigurationBase.cs
@@ -92,8 +92,8 @@ public class BaseConfigurationBase
     public int? MaxInverterAcPower { get; set; }
     public string? BleApiBaseUrl { get; set; }
     public bool UseTeslaMateAsDataSource { get; set; }
-    public decimal HomeGeofenceLongitude { get; set; }
-    public decimal HomeGeofenceLatitude { get; set; }
+    public double HomeGeofenceLongitude { get; set; }
+    public double HomeGeofenceLatitude { get; set; }
     public int HomeGeofenceRadius { get; set; }
 
     public FrontendConfiguration? FrontendConfiguration { get; set; }
